Completed Multi Layer Perceptron in Ocaml
You can test it by running "ocaml nn.ml". Generating richer data and having it learn more complex tasks for testing would be a great idea, but if it can learn the trivial case, it demonstrates that the calculus and code is correct.

Note, I am yet to write updates for the bias, and the code is a little messy. Regardless, it's functional!